+++ src/RunTime.cpp
@@ -6,9 +6,15 @@
  */
 
 #include "global.h"
+#include "utils/strings.h"
 #include "RunTime.h"
 
 RunTime::RunTime() {
+    baseDir = expandHomeDirectory("~/.OpenRaider");
+    pakDir = baseDir + "/paks";
+    audioDir = baseDir + "/music";
+    dataDir = baseDir + "/data";
+
     gameIsRunning = false;
 
 #ifdef DEBUG