+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## OpenRaider (0.1.3) xythobuz <xythobuz@xythobuz.de>
 
+    [ 20141007 ]
+    * Set all RunTime vars to sensible defaults, so OpenRaider can try to start
+      without a config file (a font is still required!)
+
     [ 20140930 ]
     * Updated imgui to version 1.14 wip
     * Console input keeps keyboard focus after command input